[MRELEASE-98] avoid NPE when running project-less
git-svn-id: https://svn.eu.apache.org/repos/asf/maven/plugins/trunk@399197 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-release-plugin/src/main/java/org/apache/maven/plugins/release/phase/AbstractRewritePomsPhase.java
@@ -491,14 +491,18 @@ public abstract class AbstractRewritePomsPhase
     {
         super.clean( config );
 
-        for ( Iterator i = config.getReactorProjects().iterator(); i.hasNext(); )
+        if ( config.getReactorProjects() != null )
         {
-            MavenProject project = (MavenProject) i.next();
-
-            File file = new File( project.getFile().getParentFile(), project.getFile().getName() + "." + pomSuffix );
-            if ( file.exists() )
+            for ( Iterator i = config.getReactorProjects().iterator(); i.hasNext(); )
             {
-                file.delete();
+                MavenProject project = (MavenProject) i.next();
+
+                File file =
+                    new File( project.getFile().getParentFile(), project.getFile().getName() + "." + pomSuffix );
+                if ( file.exists() )
+                {
+                    file.delete();
+                }
             }
         }
     }
